[Refactor] 이미지 파일 webp로 변경
