+++ src/assets/datas/keywordData.ts
@@ -1,24 +1,24 @@
-import animalImage from '@assets/images/target/animal.png';
-import childImage from '@assets/images/target/child.png';
-import collegeImage from '@assets/images/target/college-student.jpg';
-import disabledImage from '@assets/images/target/Disabled.png';
-import farmingImage from '@assets/images/target/farming-and-fishing-village1.png';
-import multiculturalImage from '@assets/images/target/multicultural-family.png';
-import oldPeopleImage from '@assets/images/target/old-man.jpg';
-import overseaImage from '@assets/images/target/Overseas-Volunteer2.png';
-import womenImage from '@assets/images/target/women2.png';
+import animalImage from '@assets/images/target/animal.webp';
+import childImage from '@assets/images/target/child.webp';
+import collegeImage from '@assets/images/target/college-student.webp';
+import disabledImage from '@assets/images/target/Disabled.webp';
+import farmingImage from '@assets/images/target/farming-and-fishing-village1.webp';
+import multiculturalImage from '@assets/images/target/multicultural-family.webp';
+import oldPeopleImage from '@assets/images/target/old-man.webp';
+import overseaImage from '@assets/images/target/Overseas-Volunteer2.webp';
+import womenImage from '@assets/images/target/women2.webp';
 
-import beautyImage from '@assets/images/subject/beauty2.png';
-import bloodImage from '@assets/images/subject/blood-donation.png';
-import briquetImage from '@assets/images/subject/briquet.jpg';
-import campaignImage from '@assets/images/subject/campaign2.png';
-import digitalImage from '@assets/images/subject/digital.png';
-import educationImage from '@assets/images/subject/education.png';
-import eventImage from '@assets/images/subject/event.png';
-import floggingImage from '@assets/images/subject/flogging.png';
-import forestFireImage from '@assets/images/subject/forest-fire.png';
-import libraryImage from '@assets/images/subject/library.png';
-import talentImage from '@assets/images/subject/talent-donation.png';
+import beautyImage from '@assets/images/subject/beauty2.webp';
+import bloodImage from '@assets/images/subject/blood-donation.webp';
+import briquetImage from '@assets/images/subject/briquet.webp';
+import campaignImage from '@assets/images/subject/campaign2.webp';
+import digitalImage from '@assets/images/subject/digital.webp';
+import educationImage from '@assets/images/subject/education.webp';
+import eventImage from '@assets/images/subject/event.webp';
+import floggingImage from '@assets/images/subject/flogging.webp';
+import forestFireImage from '@assets/images/subject/forest-fire.webp';
+import libraryImage from '@assets/images/subject/library.webp';
+import talentImage from '@assets/images/subject/talent-donation.webp';
 
 const images = {
 	animal: animalImage,
